requirements.txt
dash 
pandas
numpy
plotly.express 
plotly.graph_objects 
SARIMAX
gunicorn